adk_agriAi/requirements.txt
# Google ADK Core Dependencies
google-adk>=0.1.0
google-generativeai>=0.8.0
google-auth>=2.23.4

# Firebase/Firestore Integration
firebase-admin>=6.5.0

# ML Dependencies
tensorflow>=2.15.0
numpy>=1.26.0
scikit-learn>=1.3.0
Pillow>=10.1.0

# Web Framework Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Core Python Dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0